bug fix in create_dockerfile

+++ scripts/docker/create_dockerfile.sh
@@ -16,7 +16,7 @@ modifiedslug=$3
 # Starting with template Dockerfile, create custom Dockerfile for this one project
 #modifiedslug=$(echo ${slug} | sed 's;/;.;' | tr '[:upper:]' '[:lower:]')
 customdocker=${modifiedslug}_Dockerfile
-[ -f ${customdocker}-pipe ] { cat <${customdocker}-pipe; >/dev/null; exit 0; }
+[ -f ${customdocker}-pipe ] && { cat <${customdocker}-pipe; >/dev/null; exit 0; }
 mkfifo ${customdocker}-pipe
 cp augDockerfile ${customdocker}
 sed -i "s;<IMAGE>;detector-${modifiedslug};" ${customdocker}
